Add `--run` Flag for Broker vs. Migrations
